rename the file storing shared args and argTypes
