+++ packages/html/stories/Anchors.stories.js
@@ -33,7 +33,7 @@ import {
   globalValues,
   rubberBandTypes,
   rubberBandValues,
-} from './shared/argTypes.js';
+} from './shared/args.js';
 // style required by RubberBand
 import '@maxgraph/core/css/common.css';
