+++ packages/html/stories/Boundary.stories.js
@@ -28,7 +28,7 @@ import {
   globalValues,
   rubberBandTypes,
   rubberBandValues,
-} from './shared/argTypes.js';
+} from './shared/args.js';
 // style required by RubberBand
 import '@maxgraph/core/css/common.css';
